+++ packages/create-sitecore-jss/src/templates/nextjs/src/pages/api/editing/render.ts
@@ -1,4 +1,4 @@
-import { EditingRenderMiddleware } from '@sitecore-jss/sitecore-jss-nextjs/middleware';
+import { EditingRenderMiddleware } from '@sitecore-jss/sitecore-jss-nextjs/editing';
 
 /**
  * This Next.js API route is used to handle POST requests from Sitecore editors.
@@ -7,20 +7,21 @@ import { EditingRenderMiddleware } from '@sitecore-jss/sitecore-jss-nextjs/middl
  *
  * The `EditingRenderMiddleware` will
  *  1. Extract editing data from the Sitecore editor POST request
- *  2. Stash this data (for later use in the page render request) via the `EditingDataService`, which returns a key for retrieval
+ *  2. Stash this data (for later use in the page render request) via an `EditingDataService`, which returns a key for retrieval
  *  3. Enable Next.js Preview Mode, passing our stashed editing data key as preview data
  *  4. Invoke the actual page render request, passing along the Preview Mode cookies.
- *     This allows retrieval of the editing data in preview context (via the `EditingDataService`) - see `SitecorePagePropsFactory`
+ *     This allows retrieval of the editing data in preview context (via an `EditingDataService`) - see `SitecorePagePropsFactory`
  *  5. Return the rendered page HTML to the Sitecore editor
  */
 
-// Bump body size limit (1mb by default) for Sitecore editor payload
-// See https://nextjs.org/docs/api-routes/api-middlewares#custom-config
+// Bump body size limit (1mb by default) and disable response limit for Sitecore editor payloads
+// See https://nextjs.org/docs/api-routes/request-helpers#custom-config
 export const config = {
   api: {
     bodyParser: {
       sizeLimit: '2mb',
     },
+    responseLimit: false,
   },
 };