+++ packages/sitecore-jss/src/debug.ts
@@ -27,14 +27,14 @@ export const enableDebug = (namespaces: string) => debug.enable(namespaces);
  * Default Sitecore JSS 'debug' module debuggers. Uses namespace prefix 'sitecore-jss:'.
  * See {@link https://www.npmjs.com/package/debug} for details.
  */
-export default Object.freeze({
+export default {
   http: debug(`${rootNamespace}:http`),
   layout: debug(`${rootNamespace}:layout`),
   dictionary: debug(`${rootNamespace}:dictionary`),
-  experienceEditor: debug(`${rootNamespace}:editing`),
+  editing: debug(`${rootNamespace}:editing`),
   sitemap: debug(`${rootNamespace}:sitemap`),
   robots: debug(`${rootNamespace}:robots`),
   redirects: debug(`${rootNamespace}:redirects`),
   personalize: debug(`${rootNamespace}:personalize`),
   errorpages: debug(`${rootNamespace}:errorpages`),
-});
+};